Tidy up usage of kotlin-inject compiler
We only need to use it (and KSP) when we're creating
an actual component. We only have a few of those.
