+++ core/logging/build.gradle.kts
@@ -5,7 +5,6 @@
 plugins {
     id("app.tivi.android.library")
     id("app.tivi.kotlin.multiplatform")
-    alias(libs.plugins.ksp)
 }
 
 kotlin {
@@ -25,13 +24,6 @@ kotlin {
             }
         }
     }
-}
-
-dependencies {
-    add("kspJvm", libs.kotlininject.compiler)
-    add("kspAndroid", libs.kotlininject.compiler)
-    add("kspIosArm64", libs.kotlininject.compiler)
-    add("kspIosSimulatorArm64", libs.kotlininject.compiler)
 }
 
 android {
