+++ core/powercontroller/build.gradle.kts
@@ -5,7 +5,6 @@
 plugins {
     id("app.tivi.android.library")
     id("app.tivi.kotlin.multiplatform")
-    alias(libs.plugins.ksp)
 }
 
 kotlin {
@@ -25,10 +24,6 @@ kotlin {
             }
         }
     }
-}
-
-dependencies {
-    add("kspAndroid", libs.kotlininject.compiler)
 }
 
 android {
