+++ data/search/build.gradle.kts
@@ -4,7 +4,6 @@
 
 plugins {
     id("app.tivi.kotlin.multiplatform")
-    alias(libs.plugins.ksp)
 }
 
 kotlin {
@@ -27,8 +26,4 @@ kotlin {
             }
         }
     }
-}
-
-dependencies {
-    add("kspJvm", libs.kotlininject.compiler)
 }
