+++ ui/episode/details/build.gradle.kts
@@ -5,7 +5,6 @@
 plugins {
     id("app.tivi.android.library")
     id("app.tivi.kotlin.android")
-    alias(libs.plugins.ksp)
 }
 
 android {
@@ -38,8 +37,6 @@ dependencies {
     implementation(libs.compose.ui.tooling)
 
     implementation(libs.coil.compose)
-
-    ksp(libs.kotlininject.compiler)
 
     lintChecks(libs.slack.lint.compose)
 }
